+++ packages/types/src/generic/index.ts
@@ -2,6 +2,8 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+export * from '../ethereum';
+
 export { default as GenericAccountId } from './AccountId';
 export { default as GenericAccountIndex } from './AccountIndex';
 export { default as GenericBlock } from './Block';