+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## 1.30.0-beta.x
 
 - Adjust balance retrievals to check for `system.account` for new/old determination
-- Adjust `Address` <-> `LookupSource` definitions (no external impact, both in existence)
 - Add types for the new `proposeParachain` module (as per Rococo)
+- Adjust `Address` <-> `LookupSource` definitions (no external impact, both in existence)
+- Add Ethereum-compatible `Ethereum{AccountId, LookupSource}` types, underlying `H160`
 
 ## 1.29.1 Aug 17, 2020